Fix Couchbase 4.x index quota computation
Couchbase 4.x and higher allows an indexing
service to run on each node. When enabled
the indexing must be assigned RAM quota of at least
256MB.
This quota counts towards the total ramsize quota
(80% or available RAM).

If the indexing service is enabled compute the
service quota as 25% of the total and use the rest
for data services. If indexing is not enabled use
all available memory for data services.

Change-Id: I572e20d0755fa84e6ac902f0d10cce7a5a3daaec
Closes-Bug: DBAAS-1993