Postgres add missing argument to pg_rewind
'source-pgdata' arg is required on the pg_rewind command.
Set it to the path to the data directory
(will be the same on the master and replicas - all use
the same guest code).

Also remove the regexp-based parsing of the recovery.conf
file. Parse it with a stream_codec at the time when the file
is read.

Change-Id: I47071bf1b4f6a68aac02764c5ad6e9e535e4978b
Closes-Bug: DBAAS-1629