Ophaned Volume Not Removed on Instance Delete
When an instance is created with quota issues, the cinder volume is
created butthe nova instance fails to be created. Deleting the
instance would fail todelete the volume. Check the status of the
volume when deleting the instance, if the status of the volume is
"available" delete the volume as it is not connected to any instance.

Change-Id: If45b20fe8f5db87ff69c215c0aa0f1decac64746
Closes-Bug: DBAAS-1960