Fix MySqlEE backup & restore compression and encryption switches
The MySqlEE backup compression and encryption options are currently
hard coded as enabled in the backup and restore commands. These
settings are configurable, so the backup and restore commands are
supposed to respect the configuration settings.

This commit contain code that switches these options on/off based
on values of the backup_use_openssl_encryption and
backup_use_gzip_compression cfg settings. Unit tests for MySqlEE
has also been added.

Change-Id: Id059be0c93617cdfd20e98a5ce0bee2263784a57
Closes-Bug: dbaas-2007