Support flavor ids with leading '0'
When Trove moved to support alphanumeric flavor ids, the case of having
a numeric string id with leading zeros was not taken into account.
On the server side the code checks if the flavor id can be converted
to an int, and if so then sets it as such - the manifestation being
that if the str_id is '01' then the id will be set to '1' instead
of None as it should be.

This behavior has been fixed and unit tests created.

Change-Id: I87a8ac98d4358b75182c987f584c6098398a0c2a
Depends-On: I5acdec576a2e7da6cbfbb1cfc61c49fbbf7379af
Closes-Bug: #DBAAS-2031