Fix xtrabackup-binlog file GTID parsing
The code responsible for pulling the last GTID reference from
snapshot backup restored to the replica only handled the case
where the GTID reference contained one gtid. There are common
replication use cases where the GTID reference in the file is
comma separated list of GTID references. Change the code from
using the CSV reader to just read the entire file and split it
on tabs. This will guarantee that we get the full list of GTID
references in the file.

Change-Id: Ibbde5a4daa9741e1b997a618288bf6d469356bfe
Closes-bug: 1604914