Fix MySql replication start from incr backup
The MySql incremental restore strategy was placing a temporary
directory containing the incremental backup payload on the
instance under the MySql datadir. With Percona XtraBackup 2.2,
this was working but with the upgrade to XB 2.3, the incr.
restore from that sub-directory no longer copies over the
xtrabackup_binlog file required to configure replication. While
this is likely a bug in XB, it was thought that storing things
as sub-directories of the MySql datadir wasn't a good idea
anyway. This commit moves the incremental backup temporary
storage up one directory level (basically to the mount_point).
This fixes the issue and continues to locate the incremental
backup payload on the volume storage.

Also, augmented the replication scenario tests to test the case
where a full backup exists before replicas are configured, which
causes an incremental backup to be used for replica
provisioning.

NOTE: This only applies the "code" change from the upstream
commit as the scenario test changes weren't easily mergeable
into the dev/EE-1.7 branch.

Upstream change id is I0d5e7f37b2c376a81ed1d3c09e760fa9e6ffac8c
Closes-bug: dbaas-2047

Change-Id: I5369d4c87dbdedd43bc7ed24c02d1448cac9972a