Add icmp option for PXC, Mariadb & Oracle
A recent upstream change added a new cfg option to enable
or disable adding ICMP to the security group on a new instance.
For downstream we need to add this option to MariaDB and
the Oracles. It was also noticed that PXC was missed in the
upstream change - this will be filed/fixed upstream (but is
also fixed in this change set).

Change-Id: I4d46c4c06aaad11c052b98b0311b73b525aeea51
Closes-bug: dbaas-2088