Missing mock in test_backups
In the work done for Mysql 5.7 [1] a change was made to
detect if mysqld_safe was installed or not and then use
either mysqld or mysqld_safe as appropriate. This change
affects the code being mocked out in test_backups.
test_reset_root_password_on_mysql_restore as we now need
to mock out two restore methods instead of just one.

[1] https://review.elasticdb.org/#/c/3071/

Change-Id: I29ce69e72d5a00814f7e42deaecac0f58809629d