Fix couchbase cluster tests
Couchbase may raise an exception when busy.
The action needs to be retried.
This was failing the cluster tests.
Another issue is that the flavor for
cluster grow tests was retrieved from the
instance info which may not be initialized
when running just the cluster tests.
Skip unsupported access operation's tests.

This patch set should make Couchbase cluster
tests pass.

Change-Id: Iaa653ce189b2fa9656108d814e2e70f0f8395e4c