Fix Oracle promote-to-replica-source failure
Promotion of Oracle slave to master fails sporadically, and the
following error appears in the guest being demoted to slave:

self.app.make_read_only(read_only) ...
ORA-12514: TNS:listener does not currently know of service
requested in connect descriptor to caller

The error occurs after the "ALTER DATABASE COMMIT TO SWITCHOVER
TO STANDBY" statement when we restart the Oracle service.
The database occasionally do not register in-time to the
listener after starting up, causing the subsequent database
connection attempts to fail.

This patch change the OracleClient to make database connections
with SID instead of through the listener with a service name.
Since we no longer support multiple databases per instance,
connecting with the SID is sufficient.

Change-Id: I6c42430eb97e6fb6d557477708ea5694cd8f0fa6
Closed-Bug: dbaas-2107