Check module apply policy on create
Check module-apply permissions when creating a new instance.

If a user is not allowed to apply modules to an existing instance
he should not be allowed to create an instance with modules applied
either.

Also fix the configuration rule check on instance create.
The rule we used to check was 'configuration:update' which determines
whether a user is allowed to update an existing configuration group.
We should be checking 'instance:update' which fires when a group is attached
to an instance instead.
The reasoning is the same as with modules. If a user is not allowed to
attach a configuration group to an existing instance he should not be
allowed to create an instance with configuration group attached.

Change-Id: I71e9ebd03019681c75fc84136787551d00c86a80
Closes-Bug: DBAAS-2139