[fix gate] Cluster grow not showing locality
The scenario tests for clusters are occasionally failing when doing a
cluster-show after a cluster-grow. This seems to happen because the
locality flag on the new instances takes a few seconds (sometimes) to
populate. This results in a 'AssertionError: Cluster missing attribute
locality.' error in the cluster grow test.

The fix is to check all the instances for the server_group information
instead of just the first instance.

The cluster tests were also tweaked so they could be run separately.

Change-Id: I4f369231a8f0bc8cad338e2eb1de9cd1ca0253f6
Closes-Bug: #1621634