Fix Postgresql 9.4 replication failure
WAL streaming replication stops working and throws the following
error on the guests after promote-to-replica-source:

FATAL: could not start WAL streaming: ERROR: requested starting
point 0/4000000 on timeline 1 is not in this server's history
DETAIL: This server's history forked from timeline 1 at 0/300FDA8.

The promotion would appear to be successful because the instances'
status are set to ACTIVE despite of the error.

To address the issue, the old master must be stopped before
promoting the new slave, otherwise the timelines between the old
and the new master would diverge, causing the promotion to fail.

Also, a successful slave promotion requires a wal_level of
"hot_standby" all the time. Currently, the default wal_level in the
config.template is "minimal", and we only switch it to "hot_standby"
when we start replication.

This patch address the above issues by stopping the psql service
on the old master prior to slave promotion. It also change the
wal_level on config.template to "hot_standby". The patch also
ensure recovery mode is only disabled on the slave to be promoted,
not on an ordinary slave.

However, this patch do not address the issue with master ejection.
The issue will be addressed in a separate patch.

Change-Id: I87998ebcd9f15a8fdd1be025260d36e881faa0b3
Closes-Bug: dbaas-1019
Co-authored-by: Petr Malik <pmalik@tesora.com>