Skip 'eject valid master' replication test
This test was skipped in the scenario runs however it was
not disabled in the original api tests.  Although it
doesn't fail as often there it still happens, so the
same change has been applied to it.

See: https://review.openstack.org/#/c/368230

The test will also print out the bug number during each run
as a reminder (using the new SkipKnownBug method).

Change-Id: I931d6d72a70cc93dcd8248d9840eadf1160b9bab
Related-Bug: #1622014