Fix Oracle resize volume issue
Resizing a volume for Oracle running on rhel hangs and this error
appears in the guest log:

ExceptionPexpect: isalive() encountered condition where
"terminated" is 0, but there was no child process. Did someone
else call waitpid() on our process?

There are two calls to _unmount_volume in the
_resize_active_volume function. The second _unmount_volume call
is supposed to be a no-op if the volume is not mounted. It's
presence is to work around the undesired volume auto remount
behavior when a volume is reattached. However, it has been
discovered that the second _unmount_volume call is not always
a no-op when the volume is not mounted. As an example, the call
throws the above error for Oracle on rhel.

This commit adds a check of the current mount state of a path
before proceeding with umount. There's no point to attempt
umount on an unmounted path.

Change-Id: Icf05c76ff0a9a9b8228e3304edd4d01414359750
Closes-Bug: dbaas-1483