Fix Oracle-RA instance creation error
Oracle_RA instance creation fails with ORA-12545 error:
Connect failed because target host or object does not exist.

Oracle_RA use ezConnect to access the remote database. However,
an earlier commit removed this option, causing connectivity
issues in Oracle_RA.

This commit restores the ezConnect option on the OracleClient,
and updated the OracleVMClient to ensure that it do not use
ezConnect when making a connection to a local Oracle instance.

Change-Id: Ia2b02bcd2c9c5f9f29943cd378f7fd5f1250410e
Closes-Bug: dbaas-2233