Force delete any instance or cluster
Instances and clusters were only allowed to be force deleted when
in BUILD or ERROR state. Instances/Clusters can now be force
deleted and the status can be reset in any state.

Resetting the status will cause the Instance/Cluster to go in
ERROR state.
Force deleting an Instance/Cluster will enter a new state,
DELETING.

Change-Id: I22a0e08b4c186d1b3b6d440bdd632df73bdc539d
Closes-Bugs: DBAAS-2211, DBAAS-2227