Fix show_deleted behaviour in parameter load
A recent fix[1] changed the semantic of how "show_deleted" was used
from "return all rows including deleted" to "return only deleted".
This has caused the datastore config parameter load code to break
when faced with duplicate entries. Existing behaviour was that the
second occurence of a parameter would cause an update instead of
an insert. This commit corrects "load_parameter_by_name" to support
this behaviour.

[1] https://review.elasticdb.org/#/c/3833/

Change-Id: I298c845e08f7709ca55b901b85a576e1baa7f214
Closes-bug: dbaas-2265