Fix couchbase error handling
Couchbase add/remove had an issue with the way it handled
errors.  It would issue a rebalance command and then wait
for it to complete.  However if an error occured (either
in the execution of the command or in the rebalance itself)
the rebalance-status command would still report 'done'
(with any errors reported being ignored).  This has
been changed to wait for the actual rebalance to finish
in the task manager and re-raise any errors that occur.

The instances were also changed to not be flipped into
error, since couchbase will typically be in a sane state
even if the rebalance fails.

Unfortunately, since we don't have a 'fault' (or even
status) for clusters at the moment, the failures are not
visible.

Closes-Bug: DBAAS-2245

Change-Id: I36097182b54944fa64aa102e16e28527620a0811