Fix Postgresql 9.4 eject-replica-source issue
Although ejecting a Postgresql 9.4 master would appear to be
successful with an ACTIVE status, it actually fails behind the
scene, resulting in a replication pool without a writable
master node.

This patch corrects the issue by properly disabling the
recovery mode on the new master node. The database restart
command has also been moved to ensure all the slaves are really
connected to the new master after a promote or eject operation.

Change-Id: I759691d358556888c61c6c5a6934d0f17fc888e3
Closes-Bug: dbaas-2218