Restrict reset status command to admin
The reset status command should only be used by an Admin as
resetting the state of an instance/cluster can be dangerous.
Any user should still be able to force delete an instance/cluster.

Based on if the body of the reset command has force delete in it
or not, different policies can be used. Force delete policy when
force delete is calling the reset command and reset status when
the reset command is being called by itself.

Change-Id: I782e390f6cf9fbce346c4e510d254a1a3e1f672a
Depends-On: I236f6a97fe606c4f03f7ae3b6110d39498b76962
Partial-Bug: DBAAS-2345