Mysql 5.7 restore is failing on Ubuntu
The latest patch of Mysql 5.7 (.17) has tightened up some
security isuses. One of which appears to be that the file
provided to --log-error can't be in /tmp. While we could
change this to write to a file in a different location, it
appears that the code path that uses the error log is not
going to be used once we get to Xenial anyway (see the
comment about how restore using systemd works).

This commit removes the use of --log-error during the restore
process and all the code related to that.

Change-Id: I40a7ee5e43f915f26b78347295dc61b53eba04d5
Closes-bug: dbaas-2445