Fix DSE 5.0 backup restoration issue
trove create from a DSE 5.0 backup fails with an error in the guest,
indicating that the system_auth.credentials table is missing.

DSE 5.0 is compatible to Cassandra 3.0, and the system_auth.credentials
table has been deprecated in Cassandra 3.0. User roles and credentials
are now managed in the system_auth.roles table.

This commit implement classes for a new dse_5 datastore, which
inherit behaviors from cassandra_3.

A delay has also been implemented before initiating queries that
alters the default user's password on DSE 5.0. This delay would allow
the DSE 5.0 role manager to complete setup, preventing this error:

https://issues.apache.org/jira/browse/CASSANDRA-11942

Change-Id: I20b0b1fafb08005e2ca598773f1ba8137ebbfb32
Closes-Bug: DBAAS-2442